+++ composition-js/CHANGELOG.md
@@ -6,7 +6,15 @@ This CHANGELOG pertains only to Apollo Federation packages in the 2.x range. The
 
 > The changes noted within this `vNEXT` section have not been released yet.  New PRs and commits which introduce changes should include an entry in this `vNEXT` section as part of their development.  When a release is being prepared, a new header will be (manually) created below and the appropriate changes within that release will be moved into the new section.
 
-- _Nothing yet! Stay tuned._
+- Fix propagation of `@tag` to the supergraph and allows @tag to be repeated. Additionally, merged directives (only `@tag` and `@deprecated` currently) are not allowed on external fields anymore [PR #1592](https://github.com/apollographql/federation/pull/1592).
+
+## v2.0.0-preview.4
+
+- Released in sync with other federation packages but no changes to this package.
+
+## v2.0.0-preview.3
+
+- Released in sync with other federation packages but no changes to this package.
 
 ## v2.0.0-preview.2